{{cookiecutter.project_directory}}/requirements.dev.txt
-r requirements.lambda.txt
pytest-cov
flake8
flake8-docstrings
boto3
pytest-env
